added

+++ test1/views.py
@@ -10,6 +10,7 @@ import os
 import requests
 import tarfile
 import subprocess
+import shutil
 # Create your views here.
 
 def home(request):
@@ -65,36 +66,42 @@ def download_and_extract_tar(id, download_url, output_dir):
     tar_filename = f'{id}.tar.gz'
     tar_filepath = os.path.join(output_dir, tar_filename)
 
-    try:
-        # Download the tar.gz file
-        response = requests.get(download_url)
-        response.raise_for_status()  # Check for any download errors
+    directory_path = os.path.join(output_dir, str(id))
+    if(os.path.exists(directory_path)):
+        print("Already downloaded")
+    else:
+        try:
+            # Download the tar.gz file
+            response = requests.get(download_url)
+            response.raise_for_status()  # Check for any download errors
 
-        # Save the downloaded file
-        with open(tar_filepath, 'wb') as tar_file:
-            tar_file.write(response.content)
-            tar_file.close()
+            # Save the downloaded file
+            with open(tar_filepath, 'wb') as tar_file:
+                tar_file.write(response.content)
+                tar_file.close()
 
-        print("========== reached ==========", tar_filepath)
-        # Extract the tar.gz file
-        # with tarfile.open(tar_filepath, 'r:gz') as archive:
-        #     archive.extractall(path=output_dir)
-        result = subprocess.run(["tar", "-xvf", tar_filepath, "-C", output_dir], stdout=subprocess.PIPE, text=True)
-        print(result.stdout)
-        temp = output_dir+str(id)+'/grader/'
-        result = subprocess.run(["tar", "-xvf", temp+'grader.tar.gz', "-C", temp], stdout=subprocess.PIPE, text=True)
-        for filename in os.listdir(output_dir+str(id)+'/mounts/'):
-            temp = output_dir+str(id)+'/mounts/'
-            if filename.endswith('.tar.gz'):
-                result = subprocess.run(["tar", "-xvf", temp+filename, "-C", temp], stdout=subprocess.PIPE, text=True)
+            print("========== reached ==========", tar_filepath)
+            # Extract the tar.gz file
+            # with tarfile.open(tar_filepath, 'r:gz') as archive:
+            #     archive.extractall(path=output_dir)
+            result = subprocess.run(["tar", "-xvf", tar_filepath, "-C", output_dir], stdout=subprocess.PIPE, text=True)
+            print(result.stdout)
+            temp = output_dir+str(id)+'/grader/'
+            result = subprocess.run(["tar", "-xvf", temp+'grader.tar.gz', "-C", temp], stdout=subprocess.PIPE, text=True)
+            print(["mv", temp+'grader/*', temp])
+            result = subprocess.run(f'mv {temp}grader/* {temp}', shell=True)
+            result = subprocess.run(["rmdir" , temp+'grader/'], stdout=subprocess.PIPE, text=True)
+            # mv ./grader/* ./ && rmdir grader
+            for filename in os.listdir(output_dir+str(id)+'/mounts/'):
+                temp = output_dir+str(id)+'/mounts/'
+                if filename.endswith('.tar.gz'):
+                    result = subprocess.run(["tar", "-xvf", temp+filename, "-C", temp], stdout=subprocess.PIPE, text=True)
 
-
-
-        print(f"Downloaded and extracted '{tar_filepath}' to '{output_dir}'")
-    except requests.exceptions.RequestException as e:
-        print(f"Error: Failed to download file - {e}")
-    except tarfile.TarError as e:
-        print(f"Error: Failed to extract file - {e}")
+            print(f"Downloaded and extracted '{tar_filepath}' to '{output_dir}'")
+        except requests.exceptions.RequestException as e:
+            print(f"Error: Failed to download file - {e}")
+        except tarfile.TarError as e:
+            print(f"Error: Failed to extract file - {e}")
 
 class attemptLab(APIView):
     def get_object(self, pk):
